+++ ignite/pkg/cosmosanalysis/module/testdata/planet/x/mars/keeper/grpc_query_my_query.go
@@ -11,3 +11,7 @@ type Keeper struct{}
 func (k Keeper) MyQuery(goCtx context.Context, req *types.QueryMyQueryRequest) (*types.QueryMyQueryResponse, error) {
 	return nil, nil
 }
+
+func (k Keeper) Foo(goCtx context.Context, req *types.QueryFooRequest) (*types.QueryFooResponse, error) {
+	return nil, nil
+}