+++ changelog.md
@@ -17,6 +17,7 @@
 - nodetime built with `vercel/pkg@5.6.0`
 - Change CLI to use an events bus to print to stdout.
 - Move generated proto files to `proto/{appname}/{module}`
+- Update `pkg/cosmosanalysis` to detect when proto RPC services are using pagination.
 
 ### Fixes